+++ forgerock-core/src/main/java/org/forgerock/android/auth/AndroidPEncryptor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 ForgeRock. All rights reserved.
+ * Copyright (c) 2020 - 2021 ForgeRock. All rights reserved.
  *
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -32,7 +32,8 @@ class AndroidPEncryptor extends AndroidNEncryptor {
 
     AndroidPEncryptor(Context context, @NonNull String keyAlias, KeyUpdatedListener listener) {
         super(keyAlias, listener);
-        specBuilder.setUnlockedDeviceRequired(true);
+        //Allow access the data during screen lock
+        specBuilder.setUnlockedDeviceRequired(false);
 
         if (context.getPackageManager().hasSystemFeature(FEATURE_STRONGBOX_KEYSTORE)) {
             Logger.debug(TAG, "Strong box keystore is used.");